isl_map_underlying_set: align divs before calling this function
We need to know the meaning of the variables of the set returned by
isl_map_underlying_set in order to add it back in isl_basic_map_overlying_set,
but the number and the order of the existentially quantified variables
may change during a call to isl_map_align_divs.
We therefore need to call isl_map_align_divs first and use one of
the resulting basic maps as a model in our call to isl_basic_map_overlying_set.