isl_basic_map_gist: also exploit equalities in gisted set
Some inequalities in the gisted set may only be redundant
wrt the context inside the affine hull of the gisted set.
Although we did add the equalities to the (combined) context
before checking for redundancy, we did not perform any kind
of simplification in the combined context.
In particular, we did not perform any Gaussian elimination,
confusing the heuristic in isl_basic_map_constraint_is_redundant.