isl_tab_add_valid_eq: keep track of whether equality is negated
This is important during generalized basis reduction, because
this procedure requires the dual variables of some of the
equalities added to the tableau.  If the original equality get
negated then the dual of the original equality will be the
opposite of the dual of the equality that was actually added.

Without this fix, gbr could in some cases make completely
inappropriate decisions.