pilp solver: don't ignore feasibility test on context
In the initial implementation, the sample point of the context was
kept integer.  This was later changed in such a way that now the
sample point is just lexico-smallest rational value.
This means that after checking for an integer point, we rollback
until the state before we performed this check.
However, the results of the feasibility test were only checked
after the rollback, completely ignoring the actual results of
the feasibility test.