isl_basic_map_add_constraint: handle constraints obtained from basic maps
The bmap field of a constraint may contain a basic map that is bounded
by more than one constraint.  The constraint can therefore not be
added by simply taking the intersection with this bmap.
Instead, first create a basic map that exactly corresponds to the constraint
and then take the intersection with that basic map.