isl_basic_map_from_constraint: return copy of bmap in constraint if possible
Commit fa1cd4b (isl_basic_map_add_constraint: handle constraints obtained from
basic maps) fixed isl_basic_map_add_constraint to also work with constraints
obtained from iterating over the constraints of a basic map by
calling isl_basic_map_from_constraint.
However, this broke the case where a new constraint is created with
existentially quantified variables.  In this case, the divs may be
unknown and we can't compute them because that may result in a union.
So, try to detect this special case in isl_basic_map_from_constraint
and return the basic map stored inside the constraint.