basis_reduction_tab.c: use isl_tab_product to construct product tableau
Before, we would manually construct a tableau for the Cartesian
product of a set with itself.  Now, we first create a tableau
for the set and then construct the product tableau.
This should be more efficient and allows us to refactor
the code to pass in a tableau instead of a set.