isl_tab_basic_map_partial_lexopt: remove samples that are no longer useful
When we test an inequality in row_sign, it is useful to store any
samples found, so we can use them to speed up the computation
of future row signs.
However, when we return from a branch, then all the samples found
since we entered the branch all satisfy the constraints on that
branch and so they will not satisfy the constraints on any other
branch.  We therefore save the current number of samples when
we enter a branch and then discard all subsequently added samples
when we return from the branch.