isl_tab_pip.c: find_solutions: break when context becomes empty
Normally, find_solutions: always operates on a non-empty context.
However, when we discover a pure-parametric cut in the main
tableau, the result of adding this cut to the context may be
an empty context.  In this case, we need to break out off the
main loop, because most code assumes the context is non-empty.