isl_map_subtract: only add divs when needed
Before, the divs from all basic maps were aligne with each
other, resulting in all divs being added to all basic maps,
even if the arguments of the subtract operation are disjoint.
Now, divs are only added at the point where they are needed
and removed again if it turns out that they are not needed
in the difference.