isl_tab_pip.c: always set row sign of found row to isl_tab_row_neg
first_neg returns row that is obviously negative and usually
sets the sign of the row to isl_tab_row_neg.  However, if
the row happens to be obviously negative because the big parameter
has a negative coefficient, then it would fail to set the row sign,
while update_row_sign check that the row sign is effectively set
to isl_tab_row_neg.