isl_sample.c: initial_basis: set n_unbounded and n_zero
n_unbounded and n_zero are initialized to zero, but they can
be set to non-zero values in isl_tab_set_initial_basis_with_cone.
If isl_tab_sample is later called on an updated tab, but with
the old basis removed, then these values, in particular n_unbounded,
may not be valid anymore.
In particular, this happens in isl_tab_pip.c's gbr_get_sample.
Arguably, gbr_get_sample should reset the values of n_zero and n_unbounded
when it drops the basis, but it doens't hurt for initial_basis
to reset those values too.