isl_transitive_closure.c: path_along_delta: only avoid splitting off identity
Commit 39345fc (isl_transitive_closure.c: path_along_delta: try and avoid
splitting off identity) attempted to avoid splitting off the identity mapping.
However, it would only check if relaxing the length to zero would
not introduce any extra elements.  The path may have constraints
on the parameters, though, and so we may actually be removing elements.
This means that we need to check for equality with the identity relation.
Perhaps this makes the whole optimization pointless.