add isl_tab_sign_of_max
Make sure it's not called on zero or redundant constraints, though.
A zero constraint, in particular a dead column, could mistakenly be
considered to be manifestly unbounded.  Redundant constraints could
end up getting pivoted, which would destroy the invariant.
The internal code would never call sign_of_max on such constraints,
but external users may not be aware of this restriction.