isl_map_transitive_closure: check whether decomposition gives closed result
Some applications require the result to be closed, even if it is inexact.
The test for closedness can be fairly expensive, but I haven't been able
to trigger it yet, so presumably it will trigger only rarely.
Perhaps we should make the test configurable for those applications
that don't need a closed approximation.