isl_basic_map_from_constraint: only return copy of bmap on equality constraints
Commit b39d157 (isl_basic_map_from_constraint: return copy of bmap in
constraint if possible) changed isl_basic_map_from_constraint to return
a copy when this bmap consists of a single constraint in order to handle
the case where a new constraint is created with unknown existentially
quantified variables.  However, if the divs _are_ known, then the div
definition is simply thrown away.  If the single constraint is an equality,
then it may be possible to recover the definition, but if it is an inequality,
then this is not possible and in fact it does not make sense to have unknown divs.
Therefore, only return a copy on equality constraints.