mark some data structures and functions for export to language bindings
We prefer to only include function in interfaces to other languages
that have been explicitly marked for export.
In particular, we do not want to add functions that are obsolescent
and are only kept for backward compatibility, or functions that
only make sense for a C interface, or functions that are only meant
for other libraries such as barvinok that are tightly coupled to isl.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>