isl_map_power: apply Floyd-Warshall to compute powers as well
Before, Floyd-Warshall was only applied to compute transitive closure.
However, the power is essentially a transitive closure on an extended
relation, so we simply need to perform the nested call on such extended
relations.
In principle, we should be able to apply the incremental computation
to the power too, but changing the nested calls is a bit more tricky.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>