isl_transitive_closure.c: keep track of domains for Floyd-Warshall
When applying Floyd-Warshall to an isl_map, we don't care about
the domains anymore after we have grouped them.
However, when applying Floyd-Warshall to an isl_union_map, we
use the domains do initialize the grid with maps of the right
dimension.  We therefore need to order the domains according
to the group number so that the calling function has a chance
to find them.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>