isl_input.c: read_tuple: construct isl_dim instead of isl_basic_map
Since the tuples may contain affine expressions in terms of earlier
dimensions, it made sense to collect both the dimension information
and the possible equalities in a single object, i.e., an isl_basic_map.
However, when we introduce nested dimension specifications, we will
want to construct isl_dims recursively and then it makes more sense
to construct operate on isl_dims and to collect the equalities
in a separate data structure.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>