Revert "isl_input.c: read_tuple: construct isl_dim instead of isl_basic_map"
This reverts commit 6373465b50e93c0a8df20fdefe572a2eaebb8490.

It was thought that handling nested dims would be easier when
constructin isl_dims instead of isl_basic_maps during construction.
However, it turns out that using isl_basic_maps works just fine
and avoids carrying around the equalities separately.
Moreover, if we want to be able to parse integer divisions,
then we need to store their definitions somewhere and
an isl_basic_map is the most appropriate way to do so.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>