isl_map_coalesce: handle some cases of pairs of adjacent equalities
We don't want to handle all cases, because coalescing maps that
are too dissimilar may reduce the effectiveness of some decomposition
techniques during the computation of transitive closure.
Right now, we only perform such coalescing if there is exactly
one pair of adjacent equalities.  It's not clear if this is the best
heuristic, but it makes all transitive closure test cases pass.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>