isl_printer_print_pw_qpolynomial{,_fold}: always print special domains
If the domain of a piecewise quasipolynomial (fold) is the zero-dimensional
space, then we don't print the "[] -> " to avoid clutter.
However, if the domain has a name or if it has internal structure
then it should be printed explicitly as otherwise no distinction can
be made with the unnamed unstructured zero-dimensional space.

Signed-off-by: Sven Verdoolaege <sven@nestor.cs.kuleuven.be>