dataflow analysis: allow absence of "textual" order during sorting of sources
When sorting the potential sources, we would implicitly assume
that for any given pair of sources, one always appears before the other
at a certain level.  However, for more generic schedules, this may
not necessarily be the case.

To avoid that the results would depend on the order that the accesses
are given in case accesses overlap, we extend the partial order to
a (nearly) total order by taking into account the description of the
sources.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>