isl_map_is_subset: don't compute divs of minuend
The knowns divs of the minuend can be exploited, but only those
of the subtrahend are essential for the implementation.
Since the computation of divs can be expensive and since they
are just thrown away afterwards, it's probably best not to compute
the divs of the minuend explicitly.

In principle, isl_map_subtract could be changed in a similar way,
but there the divs are actually visible in the result, so it may
be worth it to compute them.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>