isl_tab_min: micro-optimization: postpone taking into account denominator
The denominator of the affine expression shouldn't have any effect
on the point where the minimum is reached.  Moreover, for non-rational
tableaus, it may be confusing for future optimizations to have a row
that may attain rational values.  It is therefore beter to multiply
in the denominator after the optimum has been computed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>