isl_map_identity: take dimension specification of map as input
Before, isl_basic_map_identity and isl_map_identity would take
the dimension specification of a set as input, but this could be
confusing for users.  Moreover, by taking the specification of a map
as input, we can now also construct canonical mappings between spaces
with the same dimension, but with different names or internal structure.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>