isl_basic_map_lexmin: turn off equality detection on non-parametric problems
Detecting equalities can be quite expensive and has mostly been shown
effective on parametric problems.  For non-parametric problems, we
are effectively computing several solutions as a preprocessing step
to computing a single solution, albeit a special one.
Until we obtain more evidence that this is useful,
it seems prudent to turn off this equality detection on non-parametric
problems.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>