change isl_basic_map_foreach_lexmin prototype
In particular, the callback now receives an isl_aff_list instead
of an isl_mat.  The use of an isl_mat exposed too much of the internal
details of isl.  On top of that, it was also very fragile as the meaning
of the columns in the matrix depended on the internal representation
of the separate basic set.  Any change in this internal representation,
including any simplification step, would change the meaning of the matrix.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>