rename *_fast_* functions to *_plain_*
Although the *_fast_* functions are certainly meant to be faster than
their unqualified alternatives, they are faster only because they are
not complete.  The "plain" qualification is hopefully better at conveying
the idea that these functions only consider the obvious cases.

We keep some of the *_fast_* names for backward compatibility.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>