isl_polynomial.c: move merge_divs to isl_local_space.c
In time, we will want to define isl_qpolynomials over isl_local_spaces.
We start by moving some of the handling of divs to isl_local_space.c
as it also useful for local spaces in general.
The merging is adapted to handle possibly unknown divs, which should
never be merged.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>