isl_map_simplify.c: remove unused normalize_constraints_in_compressed_space
Its last use was removed in d5e1701 (isl_basic_map_gist: be more aggressive
in removing constraints, Fri Apr 23 12:00:58 2010 +0200).
Presumably, it was kept because it thought it might still be useful in the
future.  However, it seems we haven't found a use for it so far and we
may be reorganizing the internals a bit in the not too distant future,
so let's just remove it for now.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>