isl_polynomial.c: use flex array member to avoid array out of bounds warning
The p field in struct isl_upoly_rec is meant to be a flexible array member,
but this was only introduced in C99 and so we were using a slightly
different construct.  However, we already assume C99 (in particular,
designated initializers) in the rest of the code and recent versions
of clang complain about possible array out of bounds accesses when
we don't use a flexible array member.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>