bound.c: verify_point: make sure spaces of compared qpolynomials are the same
The current isl_qpolynomial_is_equal doesn't care about the space
in which a qpolynomial lives and so we were allowed to be sloppy before.
Unfortunately, the current isl_qpolynomial_is_equal also completely
ignores the definitions of the integer divisions, so it will have to
be fixed.  This means we can't be sloppy anymore.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>