isl_constraint: drop constraint iteration
isl_constraint used to have two meanings, one as an iterator for
the constraints in a set or relation and one as an independent constraint.
We want to get rid of the first meaning because it allows the user
to change one object through another object.
In particular, we remove here the ability to iterate over the constraints
of a basic map through an isl_constraint object.

These function were never documented, but they were used by CLooG.
Now that CLooG has been updated not to use them, we can remove them.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>