isl_schedule: don't let users get access to internal isl_band_lists
We have been careful to keep the isl_schedule alive whenever the user
is holding on to an isl_band, but if the user only has a reference
to an isl_band_list inside the schedule, then the schedule can still
disappear.  We therefore hand out duplicates of internal lists.
These duplicates contain copies of the isl_bands and these outside
references keep the schedule alive.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>