decouple isl_aff_substitute_equalities
isl_aff_substitute_equalities was being called from within two contexts:
isl_pw_aff_intersect_domain and isl_aff_gist.
In the first case, the space of the equalities was that of the space
of the isl_pw_aff, while in the second case, the space of the equalities
was a lifted copy of the _local_ space of the isl_aff.
The actual implementation of isl_aff_substitute_equalities assumed
the second case.

Split isl_aff_substitute_equalities into two functions, one for each
of these two cases.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>