isl_stream_read_map: use affine expressions during parsing
The old code would use vectors and matrices, keeping track of
how the columns related to each other.  Now that isl has support
for affine expressions, the code can be simplified a lot by using
those instead.

We may lose some performance while parsing min/max expressions,
but if that ever becomes a problem, we can fix that inside
the library itself so that all users can benefit.

The switch to affine expressions necessitates the use of isl_maps
where we used to use isl_basic_maps and similarly for isl_qpolynomial
and isl_pw_qpolynomial.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>