isl_stream_read_set: accept parameter domains
In particular, accept sets without a tuple, e.g.,

	[N] -> { : N >= 1 }

At this moment, no distinction is made between a "parameter domain"
and a zero-dimensional set.  However, this will change in the near
future and so we need to have a different notation for these two
types of sets.  We start by allowing the new notation for parameter
domains.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>