make isl_pw_* object live in a map space
Before, the space of an isl_pw_* object was that of the cells.
Conceptually, however, they map elements from these cells to something
and so their space should be that of a map.
This is especially useful for isl_pw_multi_aff objects that will
be introduced later.  Since we want to be able to convert back and forth
between such objects and isl_maps, we need to keep track of the
complete map space.

isl_aff, isl_qpolynomial and isl_qpolynomial_fold objects now
also live in a map space, although internally, they still keep
track of the domain space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>