drop "nparam" argument from isl_{set,map}_read_from_{file,str}
The "nparam" argument is removed because it is inconsistent with
the rest of the API.
The "nparam" argument was only ever useful for reading sets
represented in the original PolyLib format.
For reading maps, it was of limited use because there was no way
of specifying the number of input dimensions.
In both the isl format and the extended PolyLib format, the number
of parameters can be read off from the input.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>