isl_set_intersect: properly handle intersection with parameter domains
isl_set_intersect currently allows one of its arguments to be a parameter
domain while the other is not.  This will probably change in future, but
while we still support this situation, we should make sure that the result
lives in the same space as that of the argument that is not a parameter domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>