isl_map_intersect: require arguments to live in the same space
Before, we would allow one of the arguments to be a parameter domain.
As pointed out by Andreas Kloeckner, this breaks some invariants
in other parts of the code, notably isl_pw_*_intersect_domain.
Since we have an isl_map_intersect_params now, we no longer need
this special case, so we remove it.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>