rename isl_pw_aff_add to isl_pw_aff_union_add
As pointed out by Andreas Kloeckner, isl_pw_aff_add behaves in a way
that is inconsistent with other functions defined over pairs of
isl_pw_affs in that it produces an isl_pw_aff that is defined on
the union of the definition domains of its arguments.
We therefore rename isl_pw_aff_add to isl_pw_aff_union_add and
add a new isl_pw_aff_add that produces a result that is only
defined on the intersection of the definition domains of its arguments.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>